+++ libft.h
@@ -6,7 +6,7 @@
 /*   By: Tbouder <Tbouder@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2015/11/24 12:02:58 by tbouder           #+#    #+#             */
-/*   Updated: 2016/03/14 13:01:14 by Tbouder          ###   ########.fr       */
+/*   Updated: 2016/03/14 14:35:07 by Tbouder          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -129,7 +129,7 @@ void				ft_putstr_fd(char const *s, int fd);
 void				ft_putendl(char const *s);
 void				ft_putendl_fd(char const *s, int fd);
 void				ft_putnbr(int n);
-void				ft_putnbr_base(int nb, int base, int maj);
+void				ft_putnbr_base(long nb, int base, int maj);
 void				ft_putnbr_fd(int n, int fd);
 void				ft_nbrendl(int s);
